backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
psycopg2-binary==2.9.9
redis==5.0.0

# HTTP client
httpx==0.27.0

# Scheduler
APScheduler==3.10.4

# Data processing
pandas==2.2.0
numpy==1.26.0

# ML (Phase 3)
scikit-learn==1.4.0

# Config & validation
pydantic==2.6.0
pydantic-settings==2.2.0
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0

# XML parsing
lxml==5.1.0
